Merged PR 338: Update parish table migration with new primary key column
This PR updates the existing configuration and existing migrations to use a new foreign key.

I've linked this to #963 since originally this was going to be a PR to allow the parish data to be customisable in the app settings, however this was not possible so this idea is being abandoned in favour of a known convention.

Related work items: #963