Merged PR 382: Infrastructure module
The infrastructure module allows infrastructure items to be searched for, created or edited.

Items can be searched for using a number of different parameters including ID number, parish and type.

Creating or editing an item allows users to pick a location on a map and fill in some information about the item. When editing, users also have the option to include media items.

**Still to do in a future update:**
- Search by condition or material
- Create/edit - add in condition and material options

Related work items: #966, #967, #968, #969, #997