Merged PR 426: Add DMMO module
This PR adds the first iteration of the DMMO module.

It adds all the basic functionality for DMMOs, but is not complete as we are still awaiting feedback from the customers on how to manage certain parts of the DMMO.

This is a large PR that covers lots of files. It will be deployed to staging for proper testing.

On top of DMMO changes, a number of other changes that affect other parts of the system have been done.

- Add new geometry type option in map editing for Line+RoutePicker
- Add helper for getting nearest Right of Way
- Rework error message handling in media
- Made contacts component reusable
- Add 'AddressesFromMap' component
- Add 'RouteSelectorMap' component

Related work items: #818, #820, #822