Merged PR 449: Make logout button work
This PR makes the logout button actually work, and adds a 'signed out' page so that users aren't just bumped back to sign in after signing out.

To do this, the Microsoft Identity UI had to be added, which required adding RazorPages and Controllers With Views to the Program.cs. This shouldn't cause a problem and is only used for the identity side of things.