Merged PR 453: Add user role management
This PR allows administrator users to assign user roles. It uses a very simple interface and form and uses the Beta Microsoft Graph endpoint, as the standard one doesn't seem to return email addresses?

The management interface can be accessed via the users name in the top right. The management interface also allows access to the audit log.

This PR also fixes the breadcrumbs which had incorrect '/' at the start of their URLs, meaning it didn't work properly when running on a subdirectory.

Related work items: #1092