Merged PR 634: Recent maintenance jobs on home page now display if user is not in a team
This PR changes the behaviour of the recent maintenance jobs component on the home page.

Before, if a user was not assigned to a maintenance team they would get a red error message. However some users will not be assigned to a team but still need to see recent maintenance jobs (such as senior rangers).

Now if a user is not assigned to a team, they will be able to see the most recent 5 jobs across all team areas.

Related work items: #1362