Merged PR 645: Bugfix - renamed classes to fix audit logs
This PR fixes a bug where classes that were named the same were both appearing in the audit logs of PPOs and DMMOs of the same ID number.

Classes that been renamed are:
- Application > PPOApplication
- Application > DMMOApplication

I've tested locally but will put on staging for further testing. This PR includes migrations so I'll update the staging database too.

Related work items: #1331