Merged PR 656: Add maintenance create API
This PR adds a basic maintenance job creation API endpoint. It uses a simple API Key header based authentication with keys provided as a simple list in the settings (via KeyVault for prod scenarios).

The API will take a JSON representation of a simple maintenance job, and create it (if valid) and send the user a job created notification (if email address provided) and sign them up for updates (if requested).

There is also a seperate file upload endpoint, to allow an out-of-band service to attach media to an existing job. The idea being that the forms solution will create the job quickly, then attach any media in a separate process after confirming to the user their job has been logged, to prevent them having to wait for the upload.

Related work items: #1339