Merged PR 664: Add PII and profanity checking to maintenance job creation
This PR adds optional PII (Personally Identifiable Information) detection and Content Filtering (profanity checking) to the creation of maintenance jobs. This is the one area where members of the public both supply the information and can see the information.

Currently, if a user were to include PII or profanity in a submitted maintenance job via the API, this would be immediately available for consumption by public websites. This will now go through Azures PII and Content Filtering to 'star out' and PII, and if any profanity is found, the whole description is redacted for public use. Staff will still be able to see the full details and can decide how they want it handled via a simple UI in the maintenance job.

Related work items: #1449