Merged PR 671: Added Rights of Way reopening card to homepage
This PR adds a new card component to the homepage that displays closed Rights of Way (RoW) that are due or overdue for reopening. Features include:
- Display any RoWs that have a reopening date set to within a week of today
- Display any RoWs that are overdue for reopening
- If a user is assigned to a team they will only see RoWs for their area. If no team is selected (likely for a RoW officer) then all routes will be displayed

This component does not display any RoWs that don't have a reopening date set when not set to indefinitely closed. This is because there are likely to be hundreds of them in the live data.

Related work items: #1370