Merged PR 700: Add theme switcher
This PR adds a theme switcher based on the in-built Blazor bootstrap theme switcher. This does result in a slight flash of white when the app is first loaded or reloaded, but for the sake of simplicity this is an OK drawback for now. Once in the app and using Blazor enhanced navigation no flash is seen.